backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
psycopg2-binary==2.9.7
pymongo==4.5.0
python-dotenv==1.0.0
Werkzeug==2.3.7
gunicorn==21.2.0
pytest==7.4.2
pytest-flask==1.2.0
requests==2.31.0
Pillow==10.0.1
PyPDF2==3.0.1
flask-limiter
flask-jwt-extended
flask-cors
flask-migrate
flask-sqlalchemy
flask-healthz
structlog
marshmallow
webargs

